stencil-backend/requirements.txt
fastapi
uvicorn
python-multipart
reportlab
pillow>=10.0.0,<11.0.0